requirements.txt
googlesearch-python
colorama